change structure of production components and simplify
