+++ pages/manage/productions/index.vue
@@ -9,22 +9,213 @@
           </nuxt-link>
         </small>
       </h1>
-      <productions-table />
+      <div>
+        <div class="productions-container">
+          <productions
+            :selected-id="selectedProductionId"
+            :is-loading="isLoadingProductions"
+            :productions="productions"
+            @changeSelectedProductionId="handleChangeSelectedProductionId"
+          />
+          <events
+            v-if="selectedProduction && !isLoadingEvents"
+            :is-loading="isLoadingEvents"
+            :events="events"
+            :selected-production-name="selectedProduction.name"
+          />
+        </div>
+        <div class="panel-footer">
+          <pagination
+            :rows="totalItems"
+            :per-page="productionsPerPage"
+            @changePage="changePage"
+          />
+        </div>
+      </div>
     </div>
   </div>
 </template>
 
 <script>
-  import ProductionsTable from '../../../components/productions-table';
+  import Pagination from '../../../components/pagination';
+  import Productions from '../../../components/productions/productions';
+  import Events from '../../../components/productions/events';
 
   export default {
     components: {
-      ProductionsTable,
+      Pagination,
+      Productions,
+      Events,
+    },
+    data() {
+      return {
+        pagesProductions: 1,
+        productionsPerPage: 2,
+        totalItems: 0,
+
+        selectedProductionId: '',
+        isLoadingProductions: true,
+        productions: [],
+
+        isLoadingEvents: true,
+        events: [],
+      };
+    },
+    computed: {
+      selectedProduction() {
+        return this.productions.find(
+          (production) =>
+            production.production_id === this.selectedProductionId,
+        );
+      },
+    },
+    async created() {
+      // get the first page of productions
+      await this.getProductions(0, this.productionsPerPage);
+    },
+    methods: {
+      async handleChangeSelectedProductionId(id) {
+        this.selectedProductionId = id;
+        await this.getEventsInProduction(this.selected);
+      },
+      async getAllProductions(start, limit) {
+        return await this.$api.productions.find('', start, limit);
+      },
+      async getEventById(id) {
+        return await this.$api.events.findById(id);
+      },
+      async getProductions(start, limit) {
+        this.isLoadingProductions = true;
+        const {
+          member: productions,
+          totalItems,
+        } = await this.getAllProductions(start, limit);
+
+        this.pagesProductions = Math.ceil(totalItems / this.productionsPerPage);
+        this.totalItems = totalItems;
+
+        this.productions = productions;
+        this.handleChangeSelectedProductionId(
+          this.productions[0].production_id,
+        );
+        this.isLoadingProductions = false;
+      },
+      async getEventsInProduction() {
+        this.isLoadingEvents = true;
+
+        this.events = await this.$api.events.getFromIds(
+          this.selectedProduction.events,
+        );
+
+        this.isLoadingEvents = false;
+      },
+      async changePage(newPage) {
+        const start = (newPage - 1) * this.productionsPerPage;
+        await this.getProductions(start, this.productionsPerPage);
+      },
     },
   };
 </script>
 
 <style lang="scss">
+  .productions-container {
+    display: flex;
+    width: 100%;
+    font-weight: 400;
+
+    .table {
+      font-family: 'Open Sans', Helvetica, Arial, sans-serif;
+      font-size: 15px;
+      background-color: none;
+
+      thead {
+        background-color: $white;
+      }
+
+      tbody {
+        margin-top: 0.5rem;
+        background-color: $white;
+      }
+
+      tr {
+        display: inline-flex;
+        width: 100%;
+      }
+
+      th {
+        width: 100%;
+      }
+
+      td {
+        width: 100%;
+      }
+    }
+
+    .table-productions {
+      width: 39% !important;
+      margin-right: 1%;
+
+      tbody {
+        max-height: 80vh;
+        display: block;
+        overflow-y: scroll;
+        overflow-x: hidden;
+      }
+    }
+
+    .table-events {
+      width: 60% !important;
+
+      > tbody {
+        max-height: 80vh;
+        display: block;
+        overflow-y: scroll;
+        overflow-x: hidden;
+        border-left: 5px solid #fcd1cf;
+      }
+
+      .event-item {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+      }
+
+      .event-details {
+        margin-top: 1rem;
+      }
+
+      .event-details .table {
+        margin-bottom: 0;
+      }
+
+      .event-details tbody {
+        background-color: $lightgrey;
+      }
+    }
+
+    .selected {
+      background-color: lighten(
+        $color: $udb-primary-color,
+        $amount: 50%
+      ) !important;
+
+      &:hover {
+        background-color: lighten($udb-primary-color, 40%) !important;
+      }
+    }
+
+    a {
+      font-weight: 400;
+      color: #004f94 !important;
+    }
+  }
+
+  .panel-footer {
+    padding: 10px 15px;
+    background-color: #f5f5f5;
+    border-top: 1px solid #ddd;
+  }
+
   small a {
     color: #004f94;
   }
